requirements.txt
python-telegram-bot==20.7
python-docx==1.1.0
